fix: Pr comment fixed
